K-affine-spaces (KAS)

Theory / Background

Todo

Write up mathematical background here

Syntax

The following function runs KAS:

SubspaceClustering.kas — Function

kas(X::AbstractMatrix{<:Real}, d::AbstractVector{<:Integer};
    maxiters = 100,
    rng = default_rng(),
    init = [(randsubspace(rng, size(X, 1), di), zeros(size(X, 1))) for di in d])

Cluster the N data points in the D×N data matrix X into K clusters via the K-affine-spaces (KAS) algorithm with corresponding affine space dimensions d[1],...,d[K]. Output is a KASResult containing the resulting cluster assignments c[1],...,c[N], affine space basis matrices U[1],...,U[K], bias vectors b[1],...,b[K], and metadata about the algorithm run.

KAS seeks to cluster data points by their affine space by minimizing the following total cost

\[\sum_{i=1}^N \| X[:, i] - (U[c[i]] U[c[i]]' (X[:, i] - b[c[i]]) + b[c[i]]) \|_2^2\]

with respect to the cluster assignments c[1],...,c[N], affine space basis matrices U[1],...,U[K], and bias vectors b[1],...,b[K].

Keyword arguments

maxiters::Integer = 100: maximum number of iterations

rng::AbstractRNG = default_rng(): random number generator (used when reinitializing the affine space for an empty cluster)

init::AbstractVector{<:Tuple{<:AbstractMatrix{<:AbstractFloat}, <:AbstractVector{<:AbstractFloat}}} = [(randsubspace(rng, size(X, 1), di), zeros(size(X, 1))) for di in d]: vector of K initial pair of affine space basis matrices containing U[1],...,U[K] and bias vectors containing b[1],...,b[K].

See also KASResult.

The output has the following type:

SubspaceClustering.KASResult — Type

KASResult{
    TU<:AbstractVector{<:AbstractMatrix{<:AbstractFloat}},
    Tb<:AbstractVector{<:AbstractVector{<:AbstractFloat}},
    Tc<:AbstractVector{<:Integer},
    T<:Real}

The output of kas.

Fields

U::TU: vector of affine space basis matrices U[1],...,U[K]

b::Tb: vector of bias vectors b[1],...,b[K]

c::Tc: vector of cluster assignments c[1],...,c[N]

iterations::Int: number of iterations performed

totalcost::T: final value of total cost function

counts::Vector{Int}: vector of cluster sizes counts[1],...,counts[K]

converged::Bool: final convergence status

Examples

KAS with equal subspace dimensions

Todo

Write up example here

KAS with different subspace dimensions

Todo

Write up example here

KAS with reproducible random number generation

Todo

Write up example here

KAS with custom initialization

Todo

Write up example here

KAS with custom initialization coming from clusters

Todo

Write up example here where the subspace initialization is estimated from clusters